Added Graph and DpkgGraph.
The generic Graph, Node and Edge classes are the foundation for algorithms.
The DpkgGraph class and its members represent the installed packages of the
dpkg status database and the relation between the installed packages.

This commit also includes some changes with regard to error handling, logging
and python-apt patching. Furthermore the test coverage has been increased
drastically.