Reworked Node.incoming_edges and Node.outgoing_edges.
With this commit Node.incoming_edges and Node.outgoing_edges no longer rely on
the graph to initialize the cached incoming/outgoing edges data and the data is
no longer cached on the Graph but rather on the Node itself.  Furthermore the
incoming/outgoing edges data will be initialized on the Edge constructor run.