Reworked mark_deleted methods to respect Graph hierarchy.
To respect the Graph hierarchy marking an Edge as deleted typically marks
everything above the from-node of the Edge as deleted as the Nodes and Edges
can't exist anymore without their foundation.  The only exception to this rule
is made if the hierarchie isn't violated due to edges in an or-relationship.

This commit also removed the unmark_delete methods as an unmark_delete on a
Node would recursively have to unmark Edges and in case of Edges in an
or-relationship it wouldn't be clear which one(s) to unmark.
"""