Fixed nose2 start time.
nose2 took an awful lot of time before it started to tun tests.  The reason for
this was that all the DpkgGraph unittest classes initializes a DpkgGraph
instance during __init__.  Apparently nose2 instantiates all unittest classes
multiple times before it runs the tests.  Hence the slow start time.  This
commit moves the DpkgGraph initialization to a property.