Improved performance of Node.outgoing_edges_recursive by about 60%.
The performance improvement is based on a better cache strategy.  The method
now uses 3 different caches.  The static cache is used if the result is
independent of the graph state.  The default cache is used if the node isn't
affected by any graph members that have been marked deleted.  The dynamic cache
represenets the previous caching behavior.

This new cache strategy suits Purgatory well as nodes are always deleted from
the top.  The static and default caches help a lot with the lower part of the
graph as these parts rarely change.